+++ packages/form-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @siakit/form-components
 
+## 0.0.6
+
+### Patch Changes
+
+- add password strength
+
 ## 0.0.5
 
 ### Patch Changes